+++ client/src/App.css
@@ -60,10 +60,18 @@
 }
 
 .btn {
-  background-color: transparent;
+  background-color: #f2af29;
   color: white;
   font-size: 35px;
 }
+
+.button12 {
+  display:flex;
+  justify-content: space-evenly;
+  align-items: flex-start;
+  margin-left: 0px;
+}
+
 
 .App {
   font-family: 'BrandonGrotesque', helvetica, sans-serif;
@@ -71,52 +79,6 @@
   text-align: right;
 }
 
-.navContainer {
-  font-family: 'BrandonGrotesque', helvetica, sans-serif;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  padding-top: 5px;
-  padding-left: 0px;
-  background: rgb(61, 23, 0);
-  background: linear-gradient(
-    180deg,
-    rgba(61, 23, 0, 1) 2%,
-    rgba(0, 0, 0, 0.8603816526610644) 84%
-  );
-}
-.navContainer img {
-  height: 30px;
-  width: 30px;
-  padding-left: 0px;
-  padding-top: 10px;
-}
-.navbar {
-  display: flex;
-  justify-content: center;
-  list-style-type: none;
-  font-family: 'BrandonGrotesque', helvetica, sans-serif;
-  font-weight: bold;
-  margin: 0;
-  padding-right: 600px;
-}
-.navbar li:hover {
-  padding: 10px;
-  font-family: 'BrandonGrotesque', helvetica, sans-serif;
-  color: white;
-  text-decoration: none;
-  /* display: inline-block; */
-  position: relative;
-  /* bottom: 15px; */
-  border: 1px solid red;
-}
-
-a {
-  border: solid 1px green;
-  padding-right: 20px;
-  color: #f2af29;
-  text-decoration: none;
-}
 
 .form-row {
   height: 100%;
@@ -128,39 +90,18 @@ a {
 }
 
 .TopPicks {
-  /* display: flex; */
   justify-content: center;
   color: whitesmoke;
-  /* height: 275px; */
-  /* width: 100%; */
-  /* height: 25vw; */
   background-size: cover;
   background-position: center;
-  border: 1px solid red;
 }
 
-.card-img-top {
-  border: 1px solid red;
-}
-
-.cardBlue {
-  background-color: rgb(248, 153, 37);
-}
-
-.cardYellow {
-  background-color: rgb(251, 213, 0);
-}
-
-.cardRed {
-  background-color: rgb(62, 87, 183);
-}
 
 .cardText {
   display: flex;
   justify-content: center;
-  /* padding: 85px; */
   font-weight: bold;
-  border: 1px solid red;
+
   margin: 0;
 }
 
@@ -169,7 +110,7 @@ a {
   justify-content: center;
   padding: 25px;
   font-weight: bold;
-  border: 1px solid red;
+
 }
 
 .breweryBody {
@@ -179,63 +120,37 @@ a {
   grid-template-areas:
     'header header header header'
     'main right right right'
-    'footer footer footer footer';
-  /* display: grid;
-  min-height: 100vh;
-  grid-template-columns: repeat(5, 1fr);
-  grid-template-rows: repeat(5, 1fr);
-  grid-column-gap: 0px;
-  grid-row-gap: 0px; */
-  /* border: 5px solid red; */
+    'main right right right'
+  
 }
+
 .brewerytitle {
   grid-area: header;
-  /* margin: 5vh; */
-  /* border: 1px solid blue; */
+ 
 }
+
 .brewerytitle h1 {
-  /* border: 1px solid black; */
-  /* padding-top: 20rem; */
   padding-left: 35%;
 }
-/* .middlebody {
-  display: flex;
-  justify-content: ;
-  border: 10px solid black;
-} */
+
 .brewerypic {
   grid-area: 2 / 1 / 3 / 3;
   padding-left: 15%;
   max-width: 100%;
   height: auto;
-  /* top: 10%; */
-  /* border: 1px solid red; */
 }
-.brewerypic :hover {
-  /* transform: scale(1.3); */
-  -moz-box-shadow: 0 0 10px #ccc;
-  -webkit-box-shadow: 0 0 10px #ccc;
-  box-shadow: 0 0 10px #ccc;
-}
+
+
 .brewerylist {
   grid-area: 2 / 3 / 3 / 6;
-  /* padding-right: 35%; */
-  /* border: 1px solid green; */
+  
 }
+
 .brewerylist li {
   list-style-type: none;
   padding: 1rem;
   padding-right: 35%;
   padding-left: 5%;
-}
-.mapbox {
-  grid-area: footer;
-  padding-bottom: 10%;
-  /* border: 1px solid purple; */
-}
-.mapbox h1 {
-  padding-left: 35%;
-  /* border: 1px solid black; */
 }
 
 .marker-icon {
