+++ client/src/App.jsx
@@ -4,7 +4,7 @@ import { BrowserRouter, Switch, Route } from "react-router-dom";
 import MainPage from "./Pages/MainPage";
 import LandingPage from "./Pages/LandingPage";
 import BreweryPage from "./Pages/BreweryPage";
-import AboutUs from './Pages/AboutUs';
+
 
 const App = () => (
   <BrowserRouter>
@@ -12,8 +12,7 @@ const App = () => (
     <Switch>
       <Route exact path="/" component={LandingPage} />
       <Route exact path="/Home" component={MainPage} />
-      <Route exact path="/brewery/:id" component={BreweryPage} />
-      <Route exact path="/About" component={AboutUs} /> 
+      <Route exact path="/brewery/:id" component={BreweryPage} /> 
     </Switch>
   </BrowserRouter>
 );
