--- client/src/Pages/AboutUs.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-
-export default function AboutUs() {
-  return (
-    <div>
-      <h1>This page will house a AboutUs</h1>
-    </div>
-  );
-}
